+++ console/types/field/src/from_bits.rs
@@ -78,7 +78,7 @@ mod tests {
 
     type CurrentEnvironment = Console;
 
-    const ITERATIONS: u64 = 100;
+    const ITERATIONS: usize = 100;
 
     fn check_from_bits_le() -> Result<()> {
         for i in 0..ITERATIONS {
@@ -91,7 +91,7 @@ mod tests {
             assert_eq!(expected, candidate);
 
             // Add excess zero bits.
-            let candidate = vec![given_bits, vec![false; i as usize]].concat();
+            let candidate = vec![given_bits, vec![false; i]].concat();
 
             let candidate = Field::<CurrentEnvironment>::from_bits_le(&candidate)?;
             assert_eq!(expected, candidate);
@@ -111,7 +111,7 @@ mod tests {
             assert_eq!(expected, candidate);
 
             // Add excess zero bits.
-            let candidate = vec![vec![false; i as usize], given_bits].concat();
+            let candidate = vec![vec![false; i], given_bits].concat();
 
             let candidate = Field::<CurrentEnvironment>::from_bits_be(&candidate)?;
             assert_eq!(expected, candidate);
